lnst.Controller.Job: kill method checks if job is finished
If the job is already finished there's no need to attempt the rpc call
to send a signal. Instead, the method logs this information and returns
True immediately.

The patch also includes updates to where the kill method was used while
doing this check explicitly by removing the check.

Acked-by: Jan Tluka <jtluka@redhat.com>

+++ lnst/Controller/Job.py
@@ -190,6 +190,11 @@ class Job(object):
                 signal was sent successfully.
             False if an exception was raised while sending the signal.
         """
+        if self.finished:
+            logging.info("Job {} on host {} already finished, skipping kill call."
+                    .format(self._id, self._netns.hostid))
+            return True
+
         logging.info("Sending signal {} to job {} on host {}".format(signal,
                      self._id, self._netns.hostid))
         return self._netns._machine.kill(self, signal)
